Migrate autoupdate_bot_instance_report
