+++ tool/tctl/common/collection.go
@@ -1908,27 +1908,3 @@ func (c *scopedRoleAssignmentCollection) WriteText(w io.Writer, verbose bool) er
 	_, err := t.AsBuffer().WriteTo(w)
 	return trace.Wrap(err)
 }
-
-type autoUpdateBotInstanceReportCollection struct {
-	report *autoupdatev1pb.AutoUpdateBotInstanceReport
-}
-
-func (c *autoUpdateBotInstanceReportCollection) Resources() []types.Resource {
-	return []types.Resource{types.ProtoResource153ToLegacy(c.report)}
-}
-
-func (c *autoUpdateBotInstanceReportCollection) WriteText(w io.Writer, _ bool) error {
-	t := asciitable.MakeTable([]string{"Update Group", "Version", "Count"})
-	for groupName, groupMetrics := range c.report.GetSpec().GetGroups() {
-		if groupName == "" {
-			groupName = "<no update group>"
-		}
-		for versionName, versionMetrics := range groupMetrics.GetVersions() {
-			t.AddRow([]string{groupName, versionName, strconv.Itoa(int(versionMetrics.Count))})
-		}
-	}
-	t.SortRowsBy([]int{0, 1}, true)
-
-	_, err := t.AsBuffer().WriteTo(w)
-	return trace.Wrap(err)
-}
